revert camera driver to original to fix camera

+++ drivers/media/platform/msm/camera_v2/sensor/cci/msm_cci.h
@@ -89,7 +89,7 @@ struct msm_camera_cci_gpio_cfg {
 };
 
 struct msm_camera_cci_i2c_read_cfg {
-	uint32_t addr;
+	uint16_t addr;
 	enum msm_camera_i2c_reg_addr_type addr_type;
 	uint8_t *data;
 	uint16_t num_byte;
